00Python Microservices Development/requirements.txt
boom
celery
coverage
flask_extension
flask_profiler
flask-webtest
graypy
msgpack
Molotov
nose
pika
pytest-flake8
rabbitmq
requests-mock
tox
Sphinx
WebTest
WSGIProxy2